criação de componentes Button e Rimas, melhoria da seleção da palavra
