+++ README.md
@@ -1,2 +1,12 @@
-# blocoderima
-um site react para fazer rimas.
+# **Bloco de Rima**
+Projeto em **React** para fazer rimas!
+
+## Veja funcionando:
+- Basta acessar [https://blocoderima.com](https://blocoderima.com/) e fazer sua rima.
+
+## para instalar:
+- `npm install`
+
+## para rodar:
+- `npm start`
+
