+++ my-app/src/App.css
@@ -4,5 +4,5 @@
   display: flex;
   flex-direction: column;
   align-items: center;
-  font-family:'Franklin Gothic Medium', 'Arial Narrow', Arial, sans-serif;
+  font-family: 'Franklin Gothic Medium', 'Arial Narrow', Arial, sans-serif;
 }
