+++ my-app/src/App.js
@@ -1,16 +1,14 @@
 import './App.css';
-import Bloco from "./components/Bloco"
+import Bloco from "./components/Bloco/Bloco"
 
 const App = () => {
 
-    return (
-      <div className="App">
-        <h1>Bloco de Rima</h1>
-        <Bloco/>
-      </div>
-    ); 
-    
+  return (
+    <div className="App">
+      <h1>Bloco de Rima</h1>
+      <Bloco />
+    </div>
+  );
 }
 
 export default App;
-
