+++ my-app/src/components/Bloco/Bloco.css
@@ -1,0 +1,17 @@
+.bloco {
+    margin: 1vw;
+    background-color: #eee;
+    margin: auto;
+    padding: 10px;
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+}
+
+textarea {
+    display: flex;
+    flex-direction: column;
+    max-width: 300px;
+    min-width: 300px;  
+    height: 20vh;
+}
