+++ my-app/src/components/Button/Button.css
@@ -1,30 +1,12 @@
-.bloco {
-    margin: 1vw;
-    background-color: #eee;
-    margin: auto;
-    padding: 10px;
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-}
-
-textarea {
-    display: flex;
-    flex-direction: column;
-    max-width: 300px;
-    min-width: 300px;  
-    height: 20vh;
-}
-
 input[type="button"] {
-    font-family:'Franklin Gothic Medium', 'Arial Narrow', Arial, sans-serif;
+    font-family: 'Franklin Gothic Medium', 'Arial Narrow', Arial, sans-serif;
     margin: 10px;
     max-width: 300px;
     min-width: 300px;
     width: 40vw;
     height: 60px;
     font-size: medium;
-    font-weight:700;
+    font-weight: 700;
     background-color: #00ACF1;
     border: none;
     border-radius: 10px;
@@ -37,14 +19,9 @@ input[type="button"]:active {
     width: 40vw;
     height: 60px;
     font-size: medium;
-    font-weight:700;
+    font-weight: 700;
     background-color: #0080B3;
     border: none;
     border-radius: 10px;
     transition: all 0.2s;
-}
-
-.rima {
-    color: #11F;
-    font-style: italic;
 }
