+++ my-app/src/components/Button/Button.js
@@ -1,0 +1,13 @@
+import "./Button.css"
+
+const Button = ({ textArea, onClick }) => {
+    return (
+        <input
+            type="button"
+            value={textArea}
+            onClick={onClick}>
+        </input>
+    )
+}
+
+export default Button;
