+++ my-app/src/components/Rimas/Rimas.css
@@ -1,0 +1,4 @@
+.rima {
+    color: #11F;
+    font-style: italic;
+}
